fix(cactus-web): rename DataGrid.Table file to prevent collison that breaks the docs build
