+++ modules/cactus-web/src/DataGrid/DataGrid.mdx
@@ -5,7 +5,7 @@ menu: Components
 
 import PropsTable from 'website-src/components/PropsTable'
 import DataGrid from './DataGrid'
-import Table from './Table'
+import Table from './DataGridTable'
 import SplitButton from '../SplitButton/SplitButton'
 import ScreenSizeProvider, { ScreenSizeContext, SIZES } from '../ScreenSizeProvider/ScreenSizeProvider'
 import { Component } from 'react'
